requirements.txt
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
flask-cors==4.0.0 